+++ src/core/DebugOptionalParams.ts
@@ -1,0 +1,7 @@
+interface DebugOptionalParams {
+    flagSDK ?: boolean;
+    flagACS ?: boolean;
+    flagAttachment ?: boolean;
+}
+
+export default DebugOptionalParams